covert thistory_truncate to fuse table's method
