+++ query/src/interpreters/interpreter_table_drop.rs
@@ -21,10 +21,8 @@ use common_streams::SendableDataBlockStream;
 
 use crate::catalogs::Catalog;
 use crate::interpreters::Interpreter;
-use crate::interpreters::InterpreterFactory;
 use crate::interpreters::InterpreterPtr;
 use crate::sessions::QueryContext;
-use crate::sql::PlanParser;
 
 pub struct DropTableInterpreter {
     ctx: Arc<QueryContext>,
@@ -35,15 +33,13 @@ impl DropTableInterpreter {
     pub fn try_create(ctx: Arc<QueryContext>, plan: DropTablePlan) -> Result<InterpreterPtr> {
         Ok(Arc::new(DropTableInterpreter { ctx, plan }))
     }
-    async fn truncate_history(&self) -> Result<SendableDataBlockStream> {
-        // during execution of DROP TABLE, truncate all the historical data
-        let qry = format!(
-            "select * from fuse_truncate_history('{}', '{}', 'all')",
-            self.plan.db, self.plan.table
-        );
-        let plan = PlanParser::parse(qry.as_str(), self.ctx.clone()).await?;
-        InterpreterFactory::get(self.ctx.clone(), plan)?
-            .execute(None)
+
+    async fn truncate_history(&self) -> Result<()> {
+        let db_name = self.plan.db.as_str();
+        let tbl_name = self.plan.table.as_str();
+        let tbl = self.ctx.get_table(db_name, tbl_name).await?;
+        let keep_last_snapshot = false;
+        tbl.truncate_history(self.ctx.clone(), keep_last_snapshot)
             .await
     }
 }
@@ -58,7 +54,7 @@ impl Interpreter for DropTableInterpreter {
         &self,
         _input_stream: Option<SendableDataBlockStream>,
     ) -> Result<SendableDataBlockStream> {
-        let _ = self.truncate_history().await;
+        let _ = self.truncate_history();
         let catalog = self.ctx.get_catalog();
         catalog.drop_table(self.plan.clone().into()).await?;
         Ok(Box::pin(DataBlockStream::create(
