+++ query/src/storages/fuse/table.rs
@@ -116,6 +116,14 @@ impl Table for FuseTable {
     ) -> Result<()> {
         self.do_truncate(ctx, truncate_plan).await
     }
+
+    async fn truncate_history(
+        &self,
+        ctx: Arc<QueryContext>,
+        keep_last_snapshot: bool,
+    ) -> Result<()> {
+        self.do_truncate_history(ctx, keep_last_snapshot).await
+    }
 }
 
 impl FuseTable {
