+++ query/tests/it/storages/fuse/table_test_fixture.rs
@@ -104,6 +104,12 @@ impl TestFixture {
         }
     }
 
+    pub async fn create_default_table(&self) -> Result<()> {
+        let create_table_plan = self.default_crate_table_plan();
+        let catalog = self.ctx.get_catalog();
+        catalog.create_table(create_table_plan.into()).await
+    }
+
     pub fn gen_sample_blocks(num: u32, start: i32) -> Vec<Result<DataBlock>> {
         (0..num)
             .into_iter()
