+++ tests/suites/0_stateless/09_0007_func_fuse_truncate_history.result
@@ -1,10 +1,2 @@
 3
-2	0	0
 1
-5
-6
-7
-3	5	5
-10
-1
-0
